Fix Audio Player Bug

+++ Squirrel-Engine/include/RD_Interface.h
@@ -28,7 +28,7 @@ public:
 	void updateRenderEngineOptions(std::string parameterName, float parameter);
 	void updateRenderEngineOptions(std::string parameterName, std::string parameter);
 
-	int frameCounter = 0;
+	long long frameCounter = 0;
 private:
 	Fur fur;
 };
